Book Page Count, UserBookStatus improvements
- pageCount for Books
- Exceptions for UserBookStatus
- updateRating
- updateReview
- updateFavorite

+++ src/main/java/com/timskiproekt/pagepicks/service/UserBookStatusService.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -20,28 +21,79 @@ public class UserBookStatusService {
         this.userBookStatusRepository = userBookStatusRepository;
     }
 
-    public List<UserBookStatus> getUserBookStatusesByUserId(Long userId) {
+    public List<UserBookStatus> getUserBookStatusesByUserId(Integer userId) {
         return userBookStatusRepository.findByUserId(userId);
     }
 
     public UserBookStatus saveUserBookStatus(UserBookStatus userBookStatus) {
-        return userBookStatusRepository.save(userBookStatus);
+        Optional<UserBookStatus> existingStatus = userBookStatusRepository.findByUserIdAndBookIsbn(userBookStatus.getUser().getId(), userBookStatus.getBook().getIsbn());
+        if (existingStatus.isPresent()){
+            UserBookStatus existingUserBookStatus = existingStatus.get();
+            if(existingStatus.get().getStatus() == BookStatus.READING && userBookStatus.getStatus() == BookStatus.READING){
+                throw new IllegalArgumentException("User is already reading this book");
+            }
+            if(existingStatus.get().getStatus() == BookStatus.READ && userBookStatus.getStatus() == BookStatus.READ){
+                throw new IllegalArgumentException("User has already read this book");
+            }
+            if(existingStatus.get().getStatus() == BookStatus.TO_READ && userBookStatus.getStatus() == BookStatus.TO_READ){
+                throw new IllegalArgumentException("User already has this book on their to-read list");
+            }
+            if(existingStatus.get().getStatus() != BookStatus.READING && userBookStatus.getStatus() == BookStatus.READING){
+                existingUserBookStatus.setCurrentPage(0);
+            }
+            existingUserBookStatus.setStatus(userBookStatus.getStatus());
+            return userBookStatusRepository.save(existingUserBookStatus);
+        }
+        else {
+            if(userBookStatus.getStatus() == BookStatus.READING){
+                userBookStatus.setCurrentPage(0);
+            }
+            if(userBookStatus.getStatus() == BookStatus.READ){
+                userBookStatus.setCurrentPage(userBookStatus.getBook().getPageCount());
+            }
+            return userBookStatusRepository.save(userBookStatus);
+        }
     }
 
     public void deleteUserBookStatus(Long id) {
+        if (!userBookStatusRepository.existsById(id)) {
+            throw new NoSuchElementException("User book status with id " + id + " does not exist.");
+        }
         userBookStatusRepository.deleteById(id);
     }
 
     public UserBookStatus updateCurrentPage(Long id, Integer newPage) {
-        Optional<UserBookStatus> optionalStatus = userBookStatusRepository.findById(id);
-        if (optionalStatus.isPresent()) {
-            UserBookStatus userBookStatus = optionalStatus.get();
-            if (userBookStatus.getStatus() == BookStatus.READING) {
-                userBookStatus.setCurrentPage(newPage);
-                return userBookStatusRepository.save(userBookStatus);
-            }
+        UserBookStatus userBookStatus = userBookStatusRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("UserBookStatus with id " + id + " not found"));
+        if (userBookStatus.getStatus() != BookStatus.READING) {
+            throw new IllegalArgumentException("Cannot update page number for book that is not in READING status");
         }
-        return null;
+
+        userBookStatus.setCurrentPage(newPage);
+        return userBookStatusRepository.save(userBookStatus);
     }
 
+    public UserBookStatus updateRating(Long id, Integer rating) {
+        UserBookStatus userBookStatus = userBookStatusRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("UserBookStatus with id " + id + " not found"));
+
+        userBookStatus.setRating(rating);
+        return userBookStatusRepository.save(userBookStatus);
+    }
+
+    public UserBookStatus updateReview(Long id, String review) {
+        UserBookStatus userBookStatus = userBookStatusRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("UserBookStatus with id " + id + " not found"));
+
+        userBookStatus.setReview(review);
+        return userBookStatusRepository.save(userBookStatus);
+    }
+
+    public UserBookStatus updateFavorite(Long id, boolean favorite) {
+        UserBookStatus userBookStatus = userBookStatusRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("UserBookStatus with id " + id + " not found"));
+
+        userBookStatus.setFavorite(favorite);
+        return userBookStatusRepository.save(userBookStatus);
+    }
 }
